+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 #### [Unreleased](https://github.com/siginur/SMStorage/tree/HEAD)
 [Full changelog](https://github.com/siginur/SMStorage/compare/v1.2.1...HEAD)
-- Add support for iOS 8
+- Add new StorageType: `.keychainn`
+- Add DataPolicy configuration for `files` and `keychain` storage types
+- Add get/set methods with error throwing when something went wrong
 
 ## [1.2.1](https://github.com/siginur/SMStorage/tree/v1.2.1) (2022-03-17)
 [Full changelog](https://github.com/siginur/SMStorage/compare/v1.2.0...v1.2.1)